feat: Road to the .ts. Move some elements from .js -> .ts
Update apiService for FastAPI
