--- .eslintrc.js
@@ -1,8 +1,0 @@
-module.exports = {
-  root: true,
-  extends: [
-    "plugin:vue/essential",
-    "plugin:prettier/recommended",
-    "eslint:recommended",
-  ],
-};
