+++ .gitignore
@@ -2,6 +2,7 @@
 node_modules
 db.sqlite3
 /dist
+/static
 /venv
 /static
 /tmp
