--- .husky/prepare-commit-msg
@@ -1,5 +1,0 @@
-#!/bin/sh
-. "$(dirname "$0")/_/husky.sh"
-
-make format
-exec < /dev/tty && node_modules/.bin/cz --hook || true
