+++ .prettierrc
@@ -1,0 +1,7 @@
+{
+  "semi": true,
+  "singleQuote": false,
+  "useTabs": false,
+  "trailingComma": "es5",
+  "printWidth": 80
+}
