+++ package.json
@@ -5,45 +5,49 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "start:debug": "node --inspect-brk=5858 ./src/main.js",
-    "prepare": "husky install",
-    "cm": "cz"
+    "test:unit": "vue-cli-service test:unit",
+    "test:e2e": "vue-cli-service test:e2e",
+    "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "axios": "^0.21.2",
-    "gsap": "^3.7.0",
-    "js-cookie": "^2.2.0",
-    "lodash": "^4.17.21",
-    "prettier": "^2.4.1",
-    "vue": "^2.6.12",
+    "axios": "^0.26.0",
+    "core-js": "^3.6.5",
+    "gsap": "^3.9.1",
+    "sass-loader": "^10.1.1",
+    "vue": "^2.6.11",
+    "vue-class-component": "^7.2.3",
     "vue-code-diff": "^1.2.0",
-    "vue-router": "^3.0.1",
-    "vuetify": "^2.4.11",
-    "vuex": "^3.0.1"
+    "vue-property-decorator": "^9.1.2",
+    "vue-router": "^3.2.0",
+    "vuetify": "^2.6.3",
+    "vuex": "^3.4.0"
   },
   "devDependencies": {
-    "@commitlint/cli": "^12.1.4",
-    "@commitlint/config-conventional": "^12.1.4",
-    "@vue/cli-plugin-babel": "^3.0.0-rc.12",
-    "@vue/cli-plugin-eslint": "^3.0.0-rc.12",
-    "@vue/cli-service": "^3.0.0-rc.12",
-    "commitizen": "^4.2.4",
-    "cz-conventional-changelog": "3.3.0",
-    "deepmerge": "^4.2.2",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "husky": "^7.0.1",
-    "sass": "^1.32.12",
-    "sass-loader": "^11.0.1",
-    "semantic-release": "^17.4.4",
-    "vue-cli-plugin-vuetify": "~2.4.0",
-    "vue-template-compiler": "^2.6.12"
-  },
-  "babel": {
-    "presets": [
-      "@vue/app"
-    ]
+    "@types/jest": "^24.0.19",
+    "@typescript-eslint/eslint-plugin": "^4.18.0",
+    "@typescript-eslint/parser": "^4.18.0",
+    "@vue/cli-plugin-babel": "~4.5.15",
+    "@vue/cli-plugin-e2e-cypress": "~4.5.15",
+    "@vue/cli-plugin-eslint": "~4.5.15",
+    "@vue/cli-plugin-router": "~4.5.15",
+    "@vue/cli-plugin-typescript": "~4.5.15",
+    "@vue/cli-plugin-unit-jest": "~4.5.15",
+    "@vue/cli-plugin-vuex": "~4.5.15",
+    "@vue/cli-service": "~4.5.15",
+    "@vue/eslint-config-prettier": "^6.0.0",
+    "@vue/eslint-config-typescript": "^7.0.0",
+    "@vue/test-utils": "^1.0.3",
+    "eslint": "^6.7.2",
+    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-plugin-vue": "^6.2.2",
+    "prettier": "^2.2.1",
+    "typescript": "~4.1.5",
+    "vue-cli-plugin-vuetify": "~2.4.4",
+    "vue-template-compiler": "^2.6.11",
+    "vuetify-loader": "^1.7.0",
+    "node-sass": "^6.0.1",
+    "sass": "~1.32.0",
+    "sass-loader": "^10.0.0"
   },
   "eslintConfig": {
     "root": true,
@@ -52,32 +56,33 @@
     },
     "extends": [
       "plugin:vue/essential",
-      "plugin:prettier/recommended",
-      "eslint:recommended"
+      "eslint:recommended",
+      "@vue/typescript/recommended",
+      "@vue/prettier",
+      "@vue/prettier/@typescript-eslint"
     ],
-    "rules": {},
     "parserOptions": {
-      "parser": "babel-eslint"
-    }
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
-  },
-  "release": {
-    "branches": [
-      "main"
+      "ecmaVersion": 2020
+    },
+    "rules": {},
+    "overrides": [
+      {
+        "files": [
+          "**/__tests__/*.{j,t}s?(x)",
+          "**/tests/unit/**/*.spec.{j,t}s?(x)"
+        ],
+        "env": {
+          "jest": true
+        }
+      }
     ]
   },
   "browserslist": [
     "> 1%",
     "last 2 versions",
-    "not ie <= 8"
+    "not dead"
   ],
-  "config": {
-    "commitizen": {
-      "path": "./node_modules/cz-conventional-changelog"
-    }
+  "jest": {
+    "preset": "@vue/cli-plugin-unit-jest/presets/typescript-and-babel"
   }
 }
