+++ src/components/Display.vue
@@ -299,5 +299,3 @@ export default {
   },
 };
 </script>
-
-<style></style>
