--- src/main.js
@@ -1,18 +1,0 @@
-import Vue from "vue";
-import App from "@/App.vue";
-
-import store from "@/store";
-import router from "@/router";
-
-import vuetify from "./plugins/vuetify"; // path to vuetify export
-
-Vue.config.productionTip = true;
-
-const vue = new Vue({
-  router,
-  store,
-  vuetify,
-  render: (h) => h(App),
-});
-
-vue.$mount("#app");
