+++ src/services/api.ts
@@ -1,11 +1,9 @@
 import axios from "axios";
-import Cookies from "js-cookie";
 
 export default axios.create({
   baseURL: "/api",
   timeout: 5000,
   headers: {
     "Content-Type": "application/json",
-    "X-CSRFToken": Cookies.get("csrftoken"),
   },
 });
