+++ src/store/index.ts
@@ -1,0 +1,10 @@
+import Vue from "vue";
+import Vuex from "vuex";
+
+Vue.use(Vuex);
+
+export default new Vuex.Store({
+  state: {},
+  mutations: {},
+  getters: {},
+});
